Refactored SharedState struct
